% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMeasurement-class.R
\docType{methods}
\name{as.list,EpivizMeasurement-method}
\alias{as.list,EpivizMeasurement-method}
\title{Convert \code{\link{EpivizMeasurement}} object to \code{list}}
\usage{
\S4method{as.list}{EpivizMeasurement}(x)
}
\arguments{
\item{x}{\code{\link{EpivizMeasurement}} object to coerce.}
}
\value{
a \code{list} describing measurement object
}
\description{
Convert \code{\link{EpivizMeasurement}} object to \code{list}
}

