% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseEpivizMeasurement-class.R
\name{as.list,SparseEpivizMeasurement-method}
\alias{as.list,SparseEpivizMeasurement-method}
\title{Convert \code{\link{SparseEpivizMeasurement}} object to \code{list}}
\usage{
\S4method{as.list}{SparseEpivizMeasurement}(x)
}
\arguments{
\item{x}{\code{\link{SparseEpivizMeasurement}} object to coerce.}
}
\value{
a \code{list} describing measurement object
}
\description{
Convert \code{\link{SparseEpivizMeasurement}} object to \code{list}
}
