% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseEpivizMeasurement-class.R
\name{.emptyEpivizSparseMeasurement}
\alias{.emptyEpivizSparseMeasurement}
\title{Create empty Epiviz Measurement}
\usage{
.emptyEpivizSparseMeasurement()
}
\description{
Create empty Epiviz Measurement
}
