% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizBlockData-class.R
\docType{class}
\name{EpivizBlockData-class}
\alias{EpivizBlockData}
\alias{EpivizBlockData-class}
\title{Data container for interval data.}
\description{
Used to serve data for visualizations of genomic regions only. Wraps
\code{\link{GenomicRanges}} objects.
}
\section{Methods}{

\describe{
\item{\code{get_measurements()}}{Get description of measurements served by this object}
}}
\seealso{
EpivizData
}

