% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizBpData-class.R
\docType{class}
\name{EpivizBpData-class}
\alias{EpivizBpData}
\alias{EpivizBpData-class}
\title{Container for basepair level numeric data}
\description{
Used to serve data to genomic line tracks. Wraps \code{\link{GenomicRanges}}
objects. Numeric values obtained from \code{mcols} slot.
}
\section{Methods}{

\describe{
\item{\code{get_measurements()}}{Get description of measurements served by this object}
}}
\seealso{
EpivizData
}

