% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizData-class.R
\docType{class}
\name{EpivizData-class}
\alias{EpivizData}
\alias{EpivizData-class}
\title{Data container for epiviz data server}
\description{
Data container for epiviz data server
}
\section{Methods}{

\describe{
\item{\code{update(new_object, send_request = TRUE)}}{Update underlying data object with new object}
}}

