% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizDataMgr-class.R
\docType{class}
\name{EpivizDataMgr-class}
\alias{EpivizDataMgr}
\alias{EpivizDataMgr-class}
\title{Class providing data manager for epiviz app}
\description{
Class providing data manager for epiviz app
}
\section{Methods}{

\describe{
\item{\code{is_server_closed()}}{Check if underlying server is closed, <logical>}

\item{\code{register_type(type_name, type_description)}}{Register a data type to the manager, type_name<character>, type_description<character>}
}}

