% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizFeatureData-class.R
\docType{class}
\name{EpivizFeatureData-class}
\alias{EpivizFeatureData}
\alias{EpivizFeatureData-class}
\title{Data container for RangedSummarizedExperiment objects}
\description{
Used to serve general data (used in e.g., scatter plots and heatmaps). Wraps
\code{\link{RangedSummarizedExperiment}} objects. Numeric values obtained from
\code{assays} slot
}
\section{Methods}{

\describe{
\item{\code{get_default_chart_type()}}{Get name of default chart type for this data type}

\item{\code{get_measurements()}}{Get description of measurements served by this object}
}}
\seealso{
EpivizData
}

