% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizGeneInfoData-class.R
\docType{class}
\name{EpivizGeneInfoData-class}
\alias{EpivizGeneInfoData-class}
\alias{EpivizGeneInfoData}
\title{Container for gene annotation data}
\description{
Used to serve data to gene annotation tracks. Wraps \code{\link{GenomicRanges}} objects.
Annotation obtained from columns \code{Gene} (gene symbols) and \code{Exons} (exon start and end locations).
}
\section{Methods}{

\describe{
\item{\code{get_default_chart_type()}}{Get name of default chart type for this data type}

\item{\code{get_measurements()}}{Get description of measurements served by this object}

\item{\code{get_rows(query, metadata, useOffset = FALSE)}}{Get genomic interval information overlapping query <\code{\link{GenomicRanges}}> region}
}}

\seealso{
EpivizData

register,OrganismDb
}
