% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizGuideInfoData-class.R
\docType{class}
\name{EpivizGuideInfoData-class}
\alias{EpivizGuideInfoData-class}
\alias{EpivizGuideInfoData}
\title{Container for guide annotation data}
\description{
Used to serve data to transcript annotation tracks. Wraps \code{\link{GenomicRanges}} objects.
}

\seealso{
EpivizData
}
