% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMeasurement-class.R
\docType{class}
\name{EpivizMeasurement-class}
\alias{EpivizMeasurement-class}
\alias{EpivizMeasurement}
\title{Class encapsulating a measurement description for epiviz app.}
\description{
Class encapsulating a measurement description for epiviz app.
}
