% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivizrData-package.R
\docType{package}
\name{epivizrData}
\alias{epivizrData}
\title{epivizrData}
\description{
Infrastructure package for the epivizr interactive visualization system in Bioconductor.
It provides connections between Bioconductor infrastructure objects and the epivizr
visualization framework.
}
