% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizTxInfoData-class.R
\docType{class}
\name{EpivizTxInfoData-class}
\alias{EpivizTxInfoData-class}
\alias{EpivizTxInfoData}
\title{Container for transcript annotation data}
\description{
Used to serve data to transcript annotation tracks. Wraps \code{\link{GenomicRanges}} objects.
Annotation obtained from columns \code{Transcript} (transcript_id) \code{Gene} (gene symbols) and \code{Exons} (exon start and end locations).
}

\seealso{
EpivizData

register,OrganismDb
}
