% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizMeasurement-class.R
\docType{methods}
\name{show,EpivizMeasurement-method}
\alias{show,EpivizMeasurement-method}
\title{Display measurement datasourceId and id}
\usage{
\S4method{show}{EpivizMeasurement}(object)
}
\arguments{
\item{object}{a \code{\link{EpivizMeasurement}} to display}
}
\value{
A string describing measurement
}
\description{
Display measurement datasourceId and id
}
