% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseEpivizMeasurement-class.R
\name{show,SparseEpivizMeasurement-method}
\alias{show,SparseEpivizMeasurement-method}
\title{Display measurement datasourceId and id}
\usage{
\S4method{show}{SparseEpivizMeasurement}(object)
}
\arguments{
\item{object}{a \code{\link{SparseEpivizMeasurement}} to display}
}
\value{
A string describing measurement
}
\description{
Display measurement datasourceId and id
}
