% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseEpivizMeasurement-class.R
\docType{class}
\name{SparseEpivizMeasurement-class}
\alias{SparseEpivizMeasurement-class}
\alias{SparseEpivizMeasurement}
\title{Class encapsulating a measurement description for epiviz app.}
\description{
Class encapsulating a measurement description for epiviz app.
}
