% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ahToMySQL}
\alias{ahToMySQL}
\title{Utility function to import data to a MySQL database from Annotation Hub}
\usage{
ahToMySQL(hub_ids, hub, connection, db_name, ...)
}
\arguments{
\item{hub_ids}{List of Annotation Hub record ids}

\item{hub}{Annotation Hub object}

\item{connection}{DBIConnection to a MySQL database}

\item{db_name}{Name of the database in MySQL database you are importing to}

\item{...}{arguments for toMySQL (annotation, batch)}
}
\description{
Utility function to import data to a MySQL database from Annotation Hub
}
