% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{as.data.frame,EpivizData-method}
\alias{as.data.frame,EpivizData-method}
\title{Generic as.data.frame method for EpivizData objects}
\usage{
\S4method{as.data.frame}{EpivizData}(x, query = NULL, ...)
}
\description{
Generic as.data.frame method for EpivizData objects
}
