% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizDataMgr-class.R
\docType{class}
\name{EpivizDataMgr-class}
\alias{EpivizDataMgr}
\alias{EpivizDataMgr-class}
\title{Class providing data manager for epiviz app}
\description{
Class providing data manager for epiviz app
}
\section{Methods}{

\describe{
\item{\code{add_measurements(obj, datasource_name, send_request = TRUE, ...)}}{register measurements in data manager}

\item{\code{is_ms_connected(ms_obj_or_id)}}{check if measurement object was properly added to JS app}

\item{\code{is_server_closed()}}{Check if underlying server is closed, <logical>}

\item{\code{list_measurements()}}{make a printable list of registered measurements}

\item{\code{rm_all_measurements()}}{remove all registered measurements}

\item{\code{rm_measurements(ms_obj_or_id)}}{remove registered measurments from a given data object}

\item{\code{update_measurements(ms_obj_or_id, new_object, send_request = TRUE)}}{update the underlying data object for a registered measurement (given by object or id)}
}}

