% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizDataMgr-class.R
\docType{class}
\name{EpivizDataMgr-class}
\alias{EpivizDataMgr}
\alias{EpivizDataMgr-class}
\title{Class providing data manager for epiviz app}
\description{
Class providing data manager for epiviz app
}
\section{Methods}{

\describe{
\item{\code{add_measurements(obj, datasource_name, send_request = TRUE, ...)}}{register measurements in data manager}

\item{\code{is_server_closed()}}{Check if underlying server is closed, <logical>}
}}

