% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\docType{methods}
\name{register}
\alias{register}
\alias{register,GenomicRanges-method}
\title{Generic method to register data to the data server}
\usage{
register(object, columns = NULL, ...)

\S4method{register}{GenomicRanges}(object, columns, type = c("block", "bp",
  "gene_info"), ...)
}
\arguments{
\item{object}{The object to register to data server}
}
\description{
Generic method to register data to the data server
}
\section{Methods (by class)}{
\itemize{
\item \code{GenomicRanges}: Register a 'GenomicRanges' object
}}

