% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-methods.R
\docType{methods}
\name{register}
\alias{coerceIfNeeded,GenomicRanges-method}
\alias{register}
\title{Generic method to register data to the data server}
\usage{
register(object, columns = NULL, ...)

\S4method{coerceIfNeeded}{GenomicRanges}(object, ...)
}
\arguments{
\item{object}{The object to register to data server}
}
\description{
Generic method to register data to the data server
}
\section{Methods (by class)}{
\itemize{
\item \code{GenomicRanges}: Register a 'GenomicRanges' object
}}

